Simplify Logging System, Make SQLite Singleton
Make Utils.logcat also record to the database logs if Vars.SHOUDLOG is set. No need to write code to log once to the console and once to the database. Use online suggestion of singleton sqlite to prevent leaks and make it easier to access sqlite from Utils static functions.

Rearrange log summary to see date and time by removing message preview.
Show day of the week in log details.