Overhaul the connection management to make it more reliable.
Merge AlarmReceiver into BackgroundManager as they have the same functionality: receive intents behind your back. No need to have 2 things with the same function.
Login ALWAYS runs in "async mode". NEVER EVER wait for a result from login because it will stall the ui thread and cause the app to die. Use different callback modes for login for whether a foreground (ui screen) or BackgroundManager called the login so the result can be sent to the appropriate place.
Remove convoluted 3 tiered approach to relogging in. Don't know if it really is useful but adds too many moving parts.
Remove Vars.hasInternet because there is now Utils.hasInternet() which can check on the spot. It probably is faster performance to cache the result in Vars.hasInternet but adds another moving part and another thing that can go wrong.
Remove Vars.cmdListenerRunning and Vars.dontRestart. These old strategies haven't been used in a while and are just rotting.