Handle the case when: -you're looking at user home but the login is taking a while because the connection is bad -you look at something else out of boredom -login sends a fail to broadcast_login_fg but nobody gets it because the app is in the background -it never retires until the connection dies and then comes back (hasInternet and android connectivity will start it by conincidence)
Always send the result to BackgroundManager so it can properly manage your login to the server and retry when needed. Rebroadcast the result as broadcast_login_fg so the ui can do its thing with the result too. If there is no ui waiting for it, no harm. (Not worth adding some extra moving parts for "should rebroadcast" or "broadcast mode" or some other mode switch in login async)