Close dead socket right away.
Apparently the slight timing delay in closing the sockets by going through android's async framework to close dead sockets is the cause for unrecoverable sockets dieing, the whole "java socket stupidities". (Where reconnecting to the server from an ip address previously used causes that socket to always close/die which results in a command listener dead, which results in a login, which the socket automatically closes etc... in an infinite loop).

From now on directly close the sockets by a utility function or by hand.