Prevent "misfires" from causing problems.
There appear to be cases where android will issue an "internet just reconnected" without the internet having been disconnected. The old command listener socket is still working. Attempting a relogin without closing the old connections causes some weird stuff to happen and the app's connectivity dies. On an "internet reconnected" check if command listener is null or not. If it isn't this event is a misfire. Ignore it.