Redo oIP part of VoIP.
Buffer voice by 1/3 of a second to prevent wastefully sending 40+bytes of tcp/ip/hardware headers for only 32bytes of call data. Should also reduce packet reordering correction lag.
Redo strategy to keep call in real time. For every 1/3 of a second longer it takes to download a 512byte chunk, skip that many 512byte chunks to "forward" the conversation to the present.